Change to float

+++ ch2/pr4.c
@@ -5,9 +5,9 @@
 // Yes, I know this is not a safe way of using scanf.
 int main(void)
 {
-    int amount;
+    float amount;
     printf("Enter an amount: ");
-    scanf("%d", &amount);
-    printf("With tax added: %.2f\n", amount * 1.05);
+    scanf("%f", &amount);
+    printf("With tax added: %.2f\n", amount * 1.05f);
     return 0;
 }
